Moves virtual hardware to top-level
