+++ gui/src/hw/mod.rs
