+++ gui/src/util/mod.rs
@@ -1,0 +1,3 @@
+pub use self::channel::*;
+
+mod channel;
