--- gui/src/vmm/channel/mod.rs
@@ -1,3 +1,0 @@
-pub use self::vmm::*;
-
-mod vmm;
